// fullstack_2/src/interfaces/StatsResponse.ts
export interface StatsResponse {
    sepalLengthAvg: number;
    sepalWidthAvg: number;
    petalLengthAvg: number;
    petalWidthAvg: number;
    sepalLengthMin: number;
    sepalWidthMin: number;
    petalLengthMin: number;
    petalWidthMin: number;
    sepalLengthMax: number;
    sepalWidthMax: number;
    petalLengthMax: number;
    petalWidthMax: number;
    isSuccess: boolean;
}